+++ src/mocks/rules/operation/abstract/RuleConditionalTransferLightInvariantStorage.sol
@@ -2,7 +2,7 @@
 
 pragma solidity ^0.8.20;
 // forge-lint: disable-next-line(unaliased-plain-import)
-import "OZ/token/ERC20/IERC20.sol";
+import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 
 // forge-lint: disable-next-line(unaliased-plain-import)
 import "../../validation/abstract/RuleCommonInvariantStorage.sol";